+++ react/components/Textarea/index.js
@@ -10,19 +10,27 @@ class Textarea extends Component {
     }
   }
 
+  static CharacterCountdown = props => {
+    let classes = 'mid-gray f6 mt3 '
+    if (props.value <= 10) {
+      classes += 'red'
+    }
+    return <div className={classes}>{props.value}</div>
+  }
+
   handleChange = event => {
     this.props.onChange && this.props.onChange(event)
-  };
+  }
 
   handleFocus = event => {
     this.setState({ active: true })
     this.props.onFocus && this.props.onFocus(event)
-  };
+  }
 
   handleBlur = event => {
     this.setState({ active: false })
     this.props.onBlur && this.props.onBlur(event)
-  };
+  }
 
   render() {
     const {
@@ -68,8 +76,9 @@ class Textarea extends Component {
 
     return (
       <label className="vtex-textarea">
-        {label &&
-          <span className="vtex-textarea__label db mb3 w-100">{label}</span>}
+        {label && (
+          <span className="vtex-textarea__label db mb3 w-100">{label}</span>
+        )}
         <textarea
           {...dataAttrs}
           onBlur={this.handleBlur}
@@ -94,10 +103,21 @@ class Textarea extends Component {
           {children}
         </textarea>
 
-        {errorMessage &&
-          <div className="c-danger f6 mt3 lh-title">{errorMessage}</div>}
-        {helpText &&
-          <div className="c-muted-1 f6 mt3 lh-title">{helpText}</div>}
+        <div className="flex justify-between">
+          <div>
+            {errorMessage && (
+              <div className="red f6 mt3 lh-title">{errorMessage}</div>
+            )}
+            {helpText && (
+              <div className="mid-gray f6 mt3 lh-title">{helpText}</div>
+            )}
+          </div>
+          {this.props.characterCountdown && (
+            <Textarea.CharacterCountdown
+              value={this.props.maxLength - this.props.value.length}
+            />
+          )}
+        </div>
       </label>
     )
   }
@@ -162,6 +182,8 @@ Textarea.propTypes = {
   onFocus: PropTypes.func,
   /** onBlur event */
   onBlur: PropTypes.func,
+  /** Character countdown at the bottom right */
+  characterCountdown: PropTypes.bool,
 }
 
 export default Textarea